# Conflicts:
#	src/main/java/ru/mipt/bit/platformer/actions/ActionKeyboardMapper.java
#	src/main/java/ru/mipt/bit/platformer/actions/LibGdxKeyboardMapper.java
#	src/main/java/ru/mipt/bit/platformer/event/EventAddBullet.java
#	src/main/java/ru/mipt/bit/platformer/event/EventAddTank.java
#	src/main/java/ru/mipt/bit/platformer/event/EventAddTree.java
#	src/main/java/ru/mipt/bit/platformer/graphics/objects/LibGdxGraphicBullet.java
#	src/main/java/ru/mipt/bit/platformer/graphics/objects/LibGdxGraphicObstacle.java
#	src/main/java/ru/mipt/bit/platformer/graphics/objects/LibGdxGraphicTank.java
#	src/main/java/ru/mipt/bit/platformer/models/state/Heavy.java
#	src/main/java/ru/mipt/bit/platformer/models/state/Light.java
#	src/main/java/ru/mipt/bit/platformer/models/state/Medium.java
#	src/main/java/ru/mipt/bit/platformer/models/state/TankState.java

+++ src/main/java/ru/mipt/bit/platformer/models/Colliding.java
@@ -9,7 +9,7 @@ public interface Colliding {
     /**
      * Метод занимается регистрацией полученного урона объектом
      * */
-    default void registerCollisionDamage(int collisionDamage) {}
+    default void registerCollisionDamage(float collisionDamage) {}
 
     /**
      * Метод занимается регистрацией полученного урона объектом
